{"schema":"decodex/commit/1","summary":"Split ranking text modules","authority":"manual"}
